refactor: dissolve create of ProjectController to CreateProjectController
